make test classes with all unit tests for classes methods and one class for all parameterized tests
