+++ src/test/java/FelineTest.java
@@ -1,0 +1,30 @@
+import com.example.Feline;
+import org.junit.Test;
+
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+
+public class FelineTest {
+
+    @Test
+    public void shouldBeFeline() {
+        Feline feline = new Feline();
+        String expected = "Кошачьи";
+        assertEquals(expected, feline.getFamily());
+    }
+
+    @Test
+    public void shouldBeOneKitten() throws Exception {
+        Feline feline = new Feline();
+        int expected = 1;
+        assertEquals(expected, feline.getKittens());
+    }
+
+    @Test
+    public void shouldBePredatorMeal() throws Exception {
+        Feline feline = new Feline();
+        List<String> expected = List.of("Животные", "Птицы", "Рыба");
+        assertEquals(expected, feline.eatMeat());
+    }
+}
