--- src/test/java/LionGetFoodTest.java
@@ -1,25 +1,0 @@
-import com.example.Lion;
-import com.example.Feline;
-import org.junit.Test;
-import static org.junit.Assert.assertEquals;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.MockitoJUnitRunner;
-import java.util.List;
-
-@RunWith(MockitoJUnitRunner.class)
-public class LionGetFoodTest {
-
-    @Mock
-    Feline feline;
-
-    @Test
-    public void shouldReturnFoodForPredator() throws Exception {
-        Lion lion = new Lion(feline, "Самец");
-        Mockito.when(feline.getFood("Хищник")).thenReturn(List.of("Животные", "Птицы", "Рыба"));
-        List<String> expected = List.of("Животные", "Птицы", "Рыба");
-        List<String> actual = lion.getFood();
-        assertEquals(expected, actual);
-    }
-}
